# apps/control-api/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationListResponse": {
        "properties": {
          "evaluation_count": {
            "title": "Evaluation Count",
            "type": "integer"
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRecord"
            },
            "title": "Evaluations",
            "type": "array"
          }
        },
        "required": [
          "evaluation_count"
        ],
        "title": "EvaluationListResponse",
        "type": "object"
      },
      "EvaluationRecord": {
        "properties": {
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "evaluation_id": {
            "title": "Evaluation Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "scores": {
            "$ref": "#/components/schemas/EvaluationScores"
          },
          "team": {
            "title": "Team",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "evaluation_id",
          "timestamp",
          "team",
          "model",
          "request_id",
          "decision",
          "scores"
        ],
        "title": "EvaluationRecord",
        "type": "object"
      },
...
      "GovernanceEvaluateRequest": {
        "properties": {
          "action": {
            "default": "invoke_model",
            "title": "Action",
            "type": "string"
          },
          "agent": {
            "default": "",
            "title": "Agent",
            "type": "string"
          },
          "cost_per_hour_usd": {
            "default": 0.18,
            "minimum": 0.0,
            "title": "Cost Per Hour Usd",
            "type": "number"
          },
          "cost_per_request_usd": {
            "default": 0.01,
            "minimum": 0.0,
            "title": "Cost Per Request Usd",
            "type": "number"
          },
          "environment": {
            "default": "development",
            "title": "Environment",
            "type": "string"
          },
          "forecast_monthly_cost_usd": {
            "default": 400.0,
            "minimum": 0.0,
            "title": "Forecast Monthly Cost Usd",
            "type": "number"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "title": "Groups",
            "type": "array"
          },
          "input_tokens": {
            "default": 1000,
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "default": "llama3.1:8b",
            "title": "Model",
            "type": "string"
          },
          "model_artifact_digest": {
            "default": "",
            "title": "Model Artifact Digest",
            "type": "string"
          },
          "model_revision": {
            "default": "",
            "title": "Model Revision",
            "type": "string"
          },
          "month_to_date_cost_usd": {
            "default": 100.0,
            "minimum": 0.0,
            "title": "Month To Date Cost Usd",
            "type": "number"
          },
          "namespace": {
            "default": "ai-dev",
            "title": "Namespace",
            "type": "string"
          },
          "output_tokens": {
            "default": 500,
            "minimum": 0.0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "owner": {
            "default": "alice",
            "title": "Owner",
            "type": "string"
          },
          "policy_pack": {
            "default": "",
            "title": "Policy Pack",
            "type": "string"
          },
          "prompt_text": {
            "default": "",
            "title": "Prompt Text",
            "type": "string"
          },
          "provider": {
            "default": "ollama",
            "title": "Provider",
            "type": "string"
          },
          "region": {
            "default": "",
            "title": "Region",
            "type": "string"
          },
          "requests_last_minute": {
            "default": 0,
            "minimum": 0.0,
            "title": "Requests Last Minute",
            "type": "integer"
          },
          "sensitive_data": {
            "default": false,
            "title": "Sensitive Data",
            "type": "boolean"
          },
          "subject": {
            "default": "",
            "title": "Subject",
            "type": "string"
          },
          "team": {
            "default": "platform",
            "title": "Team",
            "type": "string"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          },
          "tokens_today": {
            "default": 0,
            "minimum": 0.0,
            "title": "Tokens Today",
            "type": "integer"
          },
          "tool_access": {
            "default": false,
            "title": "Tool Access",
            "type": "boolean"
          },
          "write_permission": {
            "default": false,
            "title": "Write Permission",
            "type": "boolean"
          }
        },
        "title": "GovernanceEvaluateRequest",
        "type": "object"
      },
      "GovernanceEvaluateResponse": {
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          },
          "final_verdict": {
            "title": "Final Verdict",
            "type": "string"
          },
          "flow": {
            "items": {
              "type": "string"
            },
            "title": "Flow",
            "type": "array"
          },
          "policy_bundle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Bundle Id"
          },
          "policy_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Digest"
          },
          "policy_pack": {
            "title": "Policy Pack",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GovernanceStageResult"
            },
            "title": "Stages",
            "type": "object"
          },
          "telemetry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GovernanceTelemetryStage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "final_verdict",
          "policy_pack",
          "reasons",
          "flow",
          "stages"
        ],
        "title": "GovernanceEvaluateResponse",
        "type": "object"
      },
...
      "GovernanceStageResult": {
        "properties": {
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "factors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Factors",
            "type": "array"
          },
          "known_model": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Model"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "pack": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "risk_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Tier"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "title": "GovernanceStageResult",
        "type": "object"
      },
      "GovernanceTelemetryStage": {
        "properties": {
          "block_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Block Reasons",
            "type": "array"
          },
          "gateway_error_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Error Rate"
          },
          "gateway_p95_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway P95 Latency Ms"
          },
          "governance_block_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance Block Rate"
          },
          "prometheus_enabled": {
            "default": false,
            "title": "Prometheus Enabled",
            "type": "boolean"
          },
          "prometheus_errors": {
            "items": {
              "type": "string"
            },
            "title": "Prometheus Errors",
            "type": "array"
          },
          "quota_source": {
            "default": "request",
            "title": "Quota Source",
            "type": "string"
          },
          "requests_last_minute": {
            "default": 0,
            "title": "Requests Last Minute",
            "type": "integer"
          },
          "tenant_request_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Request Rate"
          },
          "tokens_today": {
            "default": 0,
            "title": "Tokens Today",
            "type": "integer"
          }
        },
        "title": "GovernanceTelemetryStage",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntentResolveRequest": {
        "properties": {
          "environment": {
            "default": "development",
            "title": "Environment",
            "type": "string"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "title": "Groups",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "model": {
            "default": "",
            "title": "Model",
            "type": "string"
          },
          "namespace": {
            "default": "ai-dev",
            "title": "Namespace",
            "type": "string"
          },
          "owner": {
            "default": "alice",
            "title": "Owner",
            "type": "string"
          },
          "policy_pack": {
            "default": "",
            "title": "Policy Pack",
            "type": "string"
          },
          "provider": {
            "default": "",
            "title": "Provider",
            "type": "string"
          },
          "region": {
            "default": "",
            "title": "Region",
            "type": "string"
          },
          "run_governance": {
            "default": true,
            "title": "Run Governance",
            "type": "boolean"
          },
          "subject": {
            "default": "",
            "title": "Subject",
            "type": "string"
          },
          "team": {
            "default": "platform",
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "IntentResolveRequest",
        "type": "object"
      },
      "IntentResolveResponse": {
        "properties": {
          "cluster": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IntentClusterHint"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "forbidden": {
            "default": false,
            "title": "Forbidden",
            "type": "boolean"
          },
          "governance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GovernanceEvaluateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "intent": {
            "title": "Intent",
            "type": "string"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrchestrationPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "intent",
          "confidence"
        ],
        "title": "IntentResolveResponse",
        "type": "object"
      },
...
      "ResponseEvaluateRequest": {
        "properties": {
          "cost_budget_usd": {
            "default": 0.05,
            "minimum": 0.0,
            "title": "Cost Budget Usd",
            "type": "number"
          },
          "cost_usd": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "latency_budget_ms": {
            "default": 5000.0,
            "exclusiveMinimum": 0.0,
            "title": "Latency Budget Ms",
            "type": "number"
          },
          "latency_ms": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "model": {
            "default": "llama3.1:8b",
            "title": "Model",
            "type": "string"
          },
          "prompt_text": {
            "default": "",
            "title": "Prompt Text",
            "type": "string"
          },
          "reference_context": {
            "default": "",
            "title": "Reference Context",
            "type": "string"
          },
          "request_id": {
            "default": "",
            "title": "Request Id",
            "type": "string"
          },
          "response_text": {
            "title": "Response Text",
            "type": "string"
          },
          "team": {
            "default": "platform",
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "response_text"
        ],
        "title": "ResponseEvaluateRequest",
        "type": "object"
      },
...
      "ToolEvaluateRequest": {
        "properties": {
          "action": {
            "default": "invoke",
            "title": "Action",
            "type": "string"
          },
          "agent": {
            "default": "",
            "title": "Agent",
            "type": "string"
          },
          "environment": {
            "default": "development",
            "title": "Environment",
            "type": "string"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "title": "Groups",
            "type": "array"
          },
          "mcp_server": {
            "default": "",
            "title": "Mcp Server",
            "type": "string"
          },
          "namespace": {
            "default": "ai-dev",
            "title": "Namespace",
            "type": "string"
          },
          "owner": {
            "default": "gateway",
            "title": "Owner",
            "type": "string"
          },
          "policy_pack": {
            "default": "",
            "title": "Policy Pack",
            "type": "string"
          },
          "subject": {
            "default": "",
            "title": "Subject",
            "type": "string"
          },
          "team": {
            "default": "platform",
            "title": "Team",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          },
          "write_permission": {
            "default": false,
            "title": "Write Permission",
            "type": "boolean"
          }
        },
        "required": [
          "tool"
        ],
        "title": "ToolEvaluateRequest",
        "type": "object"
      },
      "ToolEvaluateResponse": {
        "properties": {
          "final_verdict": {
            "title": "Final Verdict",
            "type": "string"
          },
          "flow": {
            "items": {
              "type": "string"
            },
            "title": "Flow",
            "type": "array"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "stages": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ToolEvaluateStageResult"
            },
            "title": "Stages",
            "type": "object"
          }
        },
        "required": [
          "final_verdict",
          "reasons",
          "flow",
          "stages"
        ],
        "title": "ToolEvaluateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control API for private AI inference infrastructure.",
    "title": "AI Infrastructure Control Plane",
    "version": "1.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluations/recent": {
      "get": {
        "operationId": "evaluations_recent_evaluations_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "team",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluations Recent",
        "tags": [
          "governance"
        ]
      }
    },
...
    "/governance/evaluate": {
      "post": {
        "operationId": "governance_evaluate_governance_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernanceEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernanceEvaluateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Governance Evaluate",
        "tags": [
          "governance"
        ]
      }
    },
    "/governance/evaluate-response": {
      "post": {
        "operationId": "governance_evaluate_response_governance_evaluate_response_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Governance Evaluate Response",
        "tags": [
          "governance"
        ]
      }
    },
    "/governance/evaluate-tool": {
      "post": {
        "operationId": "governance_evaluate_tool_governance_evaluate_tool_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolEvaluateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Governance Evaluate Tool",
        "tags": [
          "governance"
        ]
      }
    },
...
    "/intent/resolve": {
      "post": {
        "operationId": "intent_resolve_intent_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntentResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Intent Resolve",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}